Add parser generator
This commit adds an initial, but non-functional, parser generator for
TTCN-3.

A handwritten recursive descent parser for TTCN-3 works very well,
because it provides an easy way to handle various grammar issues, like
context sensitive semicolons.

However, maintaining a recursive descent parser is not cheap, because
you manually have to keep track of keywords, conflicts, workarounds,
etc.

What we do instead is a hybrid approach: We use a parser generator and
[this BNF grammar][1] to generate as much as possible (keywords, simple
rules, ...), and overwrite those parts which require manual tweaking.

[1]: https://github.com/ttcn3/specs/blob/main/ttcn3.ebnf