Add format command (#654)
This commit introduces the format command to format source files according to
the canonical TTCN-3 style.

Formatting style is a matter of taste and a subject of debate. A canonical
style settles such debates and makes code easier to read due to a consistent
style across projects and teams.

For this reason, the format command does not support custom configuration like
maximum line length or indentation style.

The canonical style is still developing and will change in the future as
is this tool. Therefore the command is still hidden. Do not use it in
production yet.