Check arguments when calling builtin functions
Until we have proper semantic analysis, we need to check the arguments
of builtin functions explicitly. This commit allows to specific formal
parameters for builtin functions. When specified the arguments will be
checked against them.

This commit also makes use of the new functionality for the existing
builtins.

Note, some error messages have been changed to be more consistent.