Add new printer interface
This commit introduces a printer interface which makes it easier switch
between tabs or spaces.

The format command now has a new option --tabs-to-spaces=N which will
convert tabs to N spaces, if specified.

Additionally we start at indentation level -1, which is a hack to
prevent module definitions being indented.