Revert test to match the old k3r behavior. (#672)
This commit reverts a test to match the old k3r version.

The test in question had been aligned with a new k3r version, which has
not been released yet. Causing local tests to fail.

Until the new k3r version is released, we revert the test to match the
old behavior.